fix: routes needs to be setup after plugins are registered for the schema validation to work
